+++ MonoDevelop/src/Main/Base/Gui/Dialogs/OptionPanels/EditTemplateDialog.cs
@@ -87,7 +87,7 @@ namespace MonoDevelop.Gui.Dialogs
 			this.VBox.PackStart(hBox2, false, false, 6);
 			
 			// set up the buttons and add them
-			this.DefaultResponse = (int) Gtk.ResponseType.Ok;
+			this.DefaultResponse = Gtk.ResponseType.Ok;
 			Gtk.Button cancelButton = new Gtk.Button(Gtk.Stock.Cancel);
 			Gtk.Button okButton = new Gtk.Button(Gtk.Stock.Ok);
 			okButton.Clicked += new EventHandler(AcceptEvent);
